+++ crates/amaru-stores/src/rocksdb/consensus/util.rs
@@ -25,7 +25,7 @@ pub(crate) const CONSENSUS_PREFIX_LEN: usize = 5;
 /// Increment this number by 1 every time the "schema" is updated, eg. a new
 /// type of keys is added, prefixes are changed, etc. then provide a migration
 /// function from the previous version.
-pub const CHAIN_DB_VERSION: u16 = 3;
+pub const CHAIN_DB_VERSION: u16 = 4;
 
 pub(crate) const ANCHOR_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"ancho";
 pub(crate) const BEST_CHAIN_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"best_";
@@ -34,6 +34,7 @@ pub(crate) const CHAIN_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"chain";
 pub(crate) const CHILD_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"child";
 pub(crate) const HEADER_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"heade";
 pub(crate) const NONCES_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"nonce";
+pub(crate) const OPCERT_PREFIX: [u8; CONSENSUS_PREFIX_LEN] = *b"ocert";
 
 /// Open a Chain DB for reading and writing.
 /// The DB _must_ exist otherwise the function will return an error.