+++ crates/amaru/src/tests/setup.rs
@@ -55,7 +55,8 @@ pub fn create_nodes(rng: &mut RandStdRng, configs: Vec<NodeTestConfig>) -> anyho
         let mut stage_graph = SimulationBuilder::default()
             .with_seed(config.seed)
             .with_mailbox_size(10000)
-            .with_trace_buffer(config.trace_buffer.clone());
+            .with_trace_buffer(config.trace_buffer.clone())
+            .with_global_epoch_offset(start_in_era().relative_time);
 
         let config = config.with_connections(connections.clone());
         let test_node_stages = create_node(&config, &mut stage_graph)?;
@@ -74,7 +75,10 @@ pub fn create_nodes(rng: &mut RandStdRng, configs: Vec<NodeTestConfig>) -> anyho
 #[allow(clippy::panic)]
 pub fn create_node(node_config: &NodeTestConfig, stage_graph: &mut impl StageGraph) -> anyhow::Result<TestNodeStages> {
     let config = node_config.make_node_configuration()?;
-    let mut global_parameters = config.global_parameters().clone();
+    let mut global_parameters: &GlobalParameters = config
+        .network
+        .as_global_parameters()
+        .unwrap_or_else(|| panic!("no default GlobalParameters for network: {}", config.network));
 
     // The chain length used when generating data is set as the `k` parameter for the node
     // in order to simulate what happens when new tips are added and trigger a move of the best
@@ -191,9 +195,15 @@ fn set_resources(node_config: &NodeTestConfig, stage_graph: &mut impl StageGraph
     stage_graph.resources().put::<ResourceHeaderStore>(node_config.chain_store.clone());
     stage_graph.resources().put::<Arc<dyn DiagnosticChainStore>>(node_config.chain_store.clone());
     stage_graph.resources().put::<ResourceBlockValidation>(block_validation.clone());
-    stage_graph.resources().put::<ResourceHasStakePools>(Arc::new(MockHasStakePools));
-    stage_graph.resources().put::<ResourceHeaderValidation>(Arc::new(MockCanValidateHeaders));
-    stage_graph.resources().put::<ResourceTxValidation>(Arc::new(MockCanValidateTxs));
+
+    #[expect(clippy::unwrap_used)]
+    let era = NetworkName::Preprod.as_era_history().unwrap();
+    #[expect(clippy::expect_used)]
+    let global = NetworkName::Preprod.as_global_parameters().cloned().expect("global parameters for preprod");
+    let cp = Arc::new(ConsensusParameters::new(global, era));
+    stage_graph.resources().put::<ResourceConsensusParameters>(cp);
+    stage_graph.resources().put::<ResourceEraHistory>(era.clone());
+    stage_graph.resources().put::<ResourcePoolSummaries>(Arc::new(PoolSummaries::default()));
     stage_graph.resources().put::<ResourceMempool<Transaction>>(node_config.mempool.clone());
     stage_graph.resources().put(node_config.connections.clone());
     Ok(())